+++ cuML/src/hmm/hmm_py.h
@@ -23,7 +23,7 @@ void init_gmmhmm_f64(HMM<double, gmm::GMM<double> > &hmm,
 void setup_gmmhmm_f64(HMM<double, gmm::GMM<double> > &hmm, double* dLlhd);
 
 void forward_backward_gmmhmm_f64(HMM<double, gmm::GMM<double> > &hmm,
-                                 double* dX, int* dlenghts, int nSeq,
+                                 double* dX, unsigned short int* dlenghts, int nSeq,
                                  bool doForward, bool doBackward, bool doGamma);
 
 void init_mhmm_f64(HMM<double, multinomial::Multinomial<double> > &hmm,
@@ -39,13 +39,13 @@ void setup_mhmm_f64(HMM<double, multinomial::Multinomial<double> > &hmm,
                     int nObs, int nSeq, double* dLlhd);
 
 void forward_backward_mhmm_f64(HMM<double, multinomial::Multinomial<double> > &hmm,
-                               int* dX, int* dlenghts, int nSeq,
+                               unsigned short int* dX, unsigned short int* dlenghts, int nSeq,
                                bool doForward, bool doBackward, bool doGamma);
 
 void viterbi_mhmm_f64(HMM<double, multinomial::Multinomial<double> > &hmm,
-                      int* dVstates, int* dX, int* dlenghts, int nSeq);
+                      unsigned short int* dVstates, unsigned short int* dX, unsigned short int* dlenghts, int nSeq);
 
 
 void m_step_mhmm_f64(HMM<double, multinomial::Multinomial<double> > &hmm,
-                     int* dX, int* dlenghts, int nSeq);
+                     unsigned short int* dX, unsigned short int* dlenghts, int nSeq);
 }
